Issue #3120797 by bramtenhove: Move access checks to a service to make it reusable for routes and blocks togethers
