+++ modules/social_features/social_event/modules/social_event_invite/social_event_invite.services.yml
@@ -1,6 +1,9 @@
 services:
   social_event_invite.access:
     class: Drupal\social_event_invite\Access\SocialEventInvitesAccess
+    arguments: ['@social_event_invite.access_helper']
+  social_event_invite.access_helper:
+    class: Drupal\social_event_invite\SocialEventInviteAccessHelper
     arguments: ['@current_route_match', '@config.factory', '@social_group.helper_service', '@entity_type.manager']
   social_event_invite.route_subscriber:
     class: Drupal\social_event_invite\Routing\RouteSubscriber
