+++ modules/social_features/social_event/modules/social_event_invite/src/Access/SocialEventInvitesAccess.php
@@ -2,15 +2,10 @@
 
 namespace Drupal\social_event_invite\Access;
 
+use Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException;
+use Drupal\Component\Plugin\Exception\PluginNotFoundException;
 use Drupal\Core\Access\AccessResult;
-use Drupal\Core\Config\ConfigFactoryInterface;
-use Drupal\Core\Entity\EntityTypeManagerInterface;
-use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
-use Drupal\Core\Routing\Access\AccessInterface;
-use Drupal\Core\Routing\RouteMatchInterface;
-use Drupal\node\NodeInterface;
-use Drupal\social_group\SocialGroupHelperService;
-use Symfony\Component\DependencyInjection\ContainerInterface;
+use Drupal\social_event_invite\SocialEventInviteAccessHelper;
 
 /**
  * Class SocialEventInvitesAccess
@@ -20,107 +15,37 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 class SocialEventInvitesAccess {
 
   /**
-   * The route match.
+   * The event invite access helper.
    *
-   * @var \Drupal\Core\Routing\RouteMatchInterface
+   * @var \Drupal\social_event_invite\SocialEventInviteAccessHelper
    */
-  protected $routeMatch;
-
-  /**
-   * Configuration factory.
-   *
-   * @var \Drupal\Core\Config\ConfigFactoryInterface
-   */
-  protected $configFactory;
-
-  /**
-   * Group helper service.
-   *
-   * @var \Drupal\social_group\SocialGroupHelperService
-   */
-  protected $groupHelperService;
-
-  /**
-   * Entity type manager.
-   *
-   * @var \Drupal\Core\Entity\EntityTypeManagerInterface
-   */
-  protected $entityTypeManager;
+  protected $accessHelper;
 
   /**
    * EventInvitesAccess constructor.
    *
-   * @param \Drupal\Core\Routing\RouteMatchInterface $routeMatch
-   *   The route match.
-   * @param \Drupal\Core\Config\ConfigFactoryInterface $configFactory
-   *   Configuration factory.
-   * @param \Drupal\social_group\SocialGroupHelperService $groupHelperService
-   *   The group helper service.
-   * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entityTypeManager
-   *   The entity type manager.
+   * @param \Drupal\social_event_invite\SocialEventInviteAccessHelper $accessHelper
+   *   The event invite access helper.
    */
-  public function __construct(RouteMatchInterface $routeMatch, ConfigFactoryInterface $configFactory, SocialGroupHelperService $groupHelperService, EntityTypeManagerInterface $entityTypeManager) {
-    $this->routeMatch = $routeMatch;
-    $this->configFactory = $configFactory;
-    $this->groupHelperService = $groupHelperService;
-    $this->entityTypeManager = $entityTypeManager;
+  public function __construct(SocialEventInviteAccessHelper $accessHelper) {
+    $this->accessHelper = $accessHelper;
   }
 
   /**
-   * Custom access check on the event invites overview.
+   * Custom access check on the invite features on events.
    *
    * @return \Drupal\Core\Access\AccessResult
-   *
-   * @throws \Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException
-   * @throws \Drupal\Component\Plugin\Exception\PluginNotFoundException
+   *   Returns the result of the access helper.
+   * @see \Drupal\social_event_invite\SocialEventInviteAccessHelper::eventFeatureAccess()
    */
-  public function access() {
-    $config = $this->configFactory->get('social_event_invite.settings');
-    $enabled_global = $config->get('invite_enroll');
-
-    // If it's globally disabled, we don't want to show the block.
-    if (!$enabled_global) {
-      return AccessResult::forbidden();
+  public function eventFeatureAccess() {
+    try {
+      return $this->accessHelper->eventFeatureAccess();
+    } catch (InvalidPluginDefinitionException $e) {
+      return AccessResult::neutral();
+    } catch (PluginNotFoundException $e) {
+      return AccessResult::neutral();
     }
-
-    // Get the group of this node.
-    $node = $this->routeMatch->getParameter('node');
-    if ($node instanceof NodeInterface) {
-      $node = $node->id();
-    }
-    $gid_from_entity = $this->groupHelperService->getGroupFromEntity([
-      'target_type' => 'node',
-      'target_id' => $node,
-    ]);
-
-    // If we have a group we need to additional checks.
-    if ($gid_from_entity !== NULL) {
-      /* @var $group \Drupal\group\Entity\GroupInterface */
-      $group = $this->entityTypeManager
-        ->getStorage('group')
-        ->load($gid_from_entity);
-
-      $enabled_for_group = $config->get('invite_group_types');
-      $enabled = FALSE;
-      if (is_array($enabled_for_group)) {
-        foreach ($enabled_for_group as $group_type) {
-          if ($group_type === $group->bundle()) {
-            $enabled = TRUE;
-            break;
-          }
-        }
-      }
-
-      // If it's not enabled for the group this event belongs to, we don't want to
-      // show the block.
-      if (!$enabled) {
-        return AccessResult::forbidden();
-      }
-    }
-
-    // If we've got this far we can be sure the user is allowed to see this
-    // block.
-    // @todo: move that function to a service.
-    return AccessResult::allowedIf(social_event_owner_or_organizer());
   }
+
 }
